Split supplier edit & products list

+++ client/src/containers/Supplier/SupplierList.js
@@ -1,31 +1,36 @@
 import React, { Component } from 'react';
+import { withRouter } from 'react-router';
 import { compose, graphql } from 'react-apollo';
-import { Table } from 'reactstrap';
+import { Button, Table } from 'reactstrap';
+import { Link } from 'react-router-dom';
 
-import SupplierListItem from './SupplierListItem';
-import SupplierAdd from './SupplierAdd';
 import ALL_SUPPLIERS_QUERY from '../../graphql/AllSuppliers.graphql';
 import UPDATE_SUPPLIER_MUTATION from '../../graphql/UpdateSupplier.graphql';
 import DELETE_SUPPLIER_MUTATION from '../../graphql/DeleteSupplier.graphql';
 
 class SupplierList extends Component {
-  _deleteItem = ({ id }) => {
+  _openDetails = ({ id }) => () =>
+    this.props.history.push(`/suppliers/${id}/edit`);
+
+  _deleteSupplier = ({ id }) => (event) => {
+    event.stopPropagation();
+
+    if (!window.confirm('Are you sure?')) { return; }
+
     this.props.deleteSupplier({
       variables: { id },
-    }).then(() => {
-      this.props.allSuppliers.refetch();
+      refetchQueries: [{query: ALL_SUPPLIERS_QUERY }],
     });
   };
 
   render() {
     const { loading, error, allSuppliers } = this.props.allSuppliers;
 
-    if (loading) {
+    if (loading && !allSuppliers) {
       return (<div>Loading</div>)
     }
 
     if (error) {
-      console.log(error);
       return (<div>An unexpected error occurred</div>)
     }
 
@@ -35,14 +40,25 @@ class SupplierList extends Component {
 
     return (
       <div>
-        <SupplierAdd createSupplier={this.props.createSupplier} />
+        <Link to="/suppliers/new">
+          <Button outline color="success">Add supplier</Button>
+        </Link>
         <Table hover>
           <tbody>
           {allSuppliers.map((supplier, index) =>
-            <SupplierListItem
-              key={index} index={index} supplier={supplier}
-              deleteItem={this._deleteItem}
-            />
+            <tr key={index} onClick={this._openDetails(supplier)}>
+              <td>{supplier.name}</td>
+              <td>{supplier.contact.first_name}</td>
+              <td>{supplier.contact.last_name}</td>
+              <td>{supplier.contact.email} { JSON.stringify(this.context.history, null, 2) } </td>
+              <td>
+                {supplier.id > 10 && // protect initial data on heroku
+                <Button outline color="danger" size="sm"
+                        onClick={this._deleteSupplier(supplier)}>
+                  Remove
+                </Button>}
+              </td>
+            </tr>
           )}
           </tbody>
         </Table>
@@ -55,4 +71,5 @@ export default compose(
   graphql(ALL_SUPPLIERS_QUERY, { name: 'allSuppliers'}),
   graphql(UPDATE_SUPPLIER_MUTATION, { name: 'updateSupplier'}),
   graphql(DELETE_SUPPLIER_MUTATION, { name: 'deleteSupplier'}),
+  withRouter,
 )(SupplierList);
